Website

+++ SecureBox/chalicelib/db.py
@@ -3,7 +3,11 @@ from boto3.dynamodb.conditions import Key, Attr
 import botocore
 import os
 from twilio.rest import Client
-from chalicelib.env import AUTH_TOKEN, ACCOUNT_SID, TWILIO_NUM
+from twilio.base.exceptions import TwilioRestException
+
+ACCOUNT_SID = os.getenv("ACCOUNT_SID")
+AUTH_TOKEN = os.getenv("AUTH_TOKEN")
+TWILIO_NUM = os.getenv("TWILIO_NUM")
 
 # DynamoDB
 dynamodb = boto3.resource('dynamodb')
@@ -15,54 +19,62 @@ class Database:
     m_table = dynamodb.Table('ClientData')
 
     def openBox(self, boxID, access):
-        try:
-            item = self.getItem(boxID)
+        item, exists = self.getItem(boxID)
+        if exists:
             if str(access) == item['access_code']:
                 return True
             if str(access) in item['orders']:
-                self.textUser(item['phone_number'], access)
+                self.textUser(item['phone_number'], item['username'], trackingID=access)
                 self.deleteOrder(boxID, access)
                 return True
-            return False
-        except botocore.exceptions.ClientError as e:
-            if e.response['Error']['Code'] == 'ResourceNotFoundException': # check this error
-                return False
-            else:
-                raise e
+        return False
 
-    def register(self, boxID, access_code, orders, lock_status, phone_number, email, password):
-        item = {
-            'box_id': int(boxID),
-            'access_code': str(access_code),
-            'orders': set(orders),
-            'locked': bool(lock_status),
-            'phone_number': str(phone_number),
-            'email': str(email),
-            'password': str(password),
-        }
+    def register(self, boxID, access_code, phone_number, username):
+        response = self.m_table.query(
+            KeyConditionExpression=Key('box_id').eq(int(boxID))
+        )
         try:
+            if (boxID == response['Items'][0]['box_id']):
+                return False
+        except IndexError:
+            item = {
+                'box_id': int(boxID),
+                'access_code': str(access_code),
+                'locked': True,
+                'phone_number': str(phone_number),
+                'username': str(username)
+            }
             self.m_table.put_item(
                 Item=item
-            )
+            )       
+            self.textUser(phone_number, username, trackingID=None, event='register') 
             return True
-        except botocore.exceptions.ClientError as e:
-            if e.response['Error']['Code'] != 'ConditionalCheckFailedException':
-                print('This box_id already exists.')
-            else:
-                raise e
-            return False
 
     def unregister(self, boxID):
         self.m_table.delete_item(
             Key = {'box_id': int(boxID)}
         )
     
-    def textUser(self, phoneNumber, trackingID):
-        client.messages.create(
-            to=phoneNumber,
-            from_=TWILIO_NUM,
-            body='Your order with tracking ID: ' + trackingID + ' has arrived. \nIt is safe with SecureBox!'
-        )
+    def textUser(self, phoneNumber, username, trackingID, event='authenticate'):
+        if event == 'register':
+            try:
+                client.messages.create(
+                    to=phoneNumber,
+                    from_=TWILIO_NUM,
+                    body='Hey ' + username + '! Thanks for registering with SecureBox. Add a new order from the home page!'
+                )
+            except TwilioRestException:
+                print('add number to twilio account')
+        else:
+            try: 
+                client.messages.create(
+                    to=phoneNumber,
+                    from_=TWILIO_NUM,
+                    body='Hey ' + username + '! Your order with tracking ID: ' + trackingID + ' has arrived. \nIt is safe with SecureBox!'
+                )
+            except TwilioRestException:
+                print('add phone number to twilio account')
+
     
     def addOrder(self, boxID, tracking_id):
         self.m_table.update_item(
@@ -78,11 +90,23 @@ class Database:
             ExpressionAttributeValues = {':order': {tracking_id}}
         )
 
+    def getOrders(self, boxID):
+        item, exists = self.getItem(boxID)
+        if exists:
+            try:
+                return item['orders']
+            except KeyError:
+                return {'no orders.'}
+        return {'invalid box ID.'}
+
     def getItem(self, boxID):
-        item = self.m_table.get_item(
-            Key={'box_id': int(boxID)}
-        )
-        return item['Item']
+        try:
+            item = self.m_table.get_item(
+                Key={'box_id': int(boxID)}
+            )
+            return item['Item'], True
+        except KeyError:
+            return 'invalid box ID.', False
 
     def setLockStatus(self, boxID, lockStatus):
         self.m_table.update_item(
@@ -92,8 +116,10 @@ class Database:
         )
     
     def getLockStatus(self, boxID):
-        item = self.getItem(int(boxID))
-        return item['locked']
+        item, exists = self.getItem(int(boxID))
+        if exists:
+            return item['locked']
+        return 'invalid box ID.'
 
     def setAccessCode(self, boxID, access_code):
         self.m_table.update_item(
@@ -103,19 +129,23 @@ class Database:
         )
     
     def getAccessCode(self, boxID):
-        item = self.getItem(int(boxID))
-        return item['access_code']
+        item, exists = self.getItem(int(boxID))
+        if exists:
+            return item['access_code']
+        return 'invalid box ID.'
 
-    def setEmail(self, boxID, email):
+    def setUsername(self, boxID, username):
         self.m_table.update_item(
             Key = {'box_id': int(boxID)},
-            UpdateExpression = 'SET email = :new_email',
-            ExpressionAttributeValues = {':new_email': email}
+            UpdateExpression = 'SET username = :username',
+            ExpressionAttributeValues = {':username': username}
         )
         
-    def getEmail(self, boxID):
-        item = self.getItem(int(boxID))
-        return item['email']
+    def getUsername(self, boxID):
+        item, exists = self.getItem(int(boxID))
+        if exists:
+            return item['username']
+        return 'invalid box ID.'
 
     def setPhoneNumber(self, boxID, phone_number):
         self.m_table.update_item(
@@ -125,6 +155,7 @@ class Database:
         )
 
     def getPhoneNumber(self, boxID):
-        item = self.getItem(int(boxID))
-        return item['phone_number']
-
+        item, exists = self.getItem(int(boxID))
+        if exists:
+            return item['phone_number']
+        return 'invalid box ID.'
